fixed some stuff
I have fixed that date thing, it was really easy tbh. Added a timeout of 10 seconds if a client says nothing to the server and made it so if the string is not date, exit or quit, the output is the string reversed. Still has some problems that im gonna fix tho